+++ src/main/kotlin/com/github/sleepypanda/feesh/FeeshMod.kt
@@ -72,6 +72,7 @@ class FeeshMod : ClientModInitializer {
         GuiUtils.init()
         PriceUtils.init()
         KeybindUtils.init()
+        RareDropAlertUtils.init()
 
         // Event publishers
         SeaCreaturesPublisher.init()